// src/assured/enums/assured-queue.enums.ts
export enum AssuredQueueAdvanceReason {
  FULL = 'FULL',
  DRIVER_CANCELLED = 'DRIVER_CANCELLED',
  DRIVER_NO_SHOW = 'DRIVER_NO_SHOW',
  FORCE_PUBLISH = 'FORCE_PUBLISH',
}

export enum AssuredQueueEventType {
  ENQUEUED = 'ENQUEUED',
  PROMOTED = 'PROMOTED',
  FULL_PROMOTION = 'FULL_PROMOTION',
  FORCE_PUBLISHED = 'FORCE_PUBLISHED',
  CANCELLED_PROMOTION = 'CANCELLED_PROMOTION',
}